summative-green-academy-api-final-phase-ccc-main/requirements.txt
asgiref==3.8.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.32.10
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
psutil==7.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.9.0
zope.event==5.0
zope.interface==7.2
python-dotenv==1.0.1
